requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9

